Free the alternate signal stack if this PAL's thread allocated it. (#19539)
Free the alternate signal stack if this PAL's thread allocated it.

Move Ensure/FreeSignalAlternateStack to CPalThread class.

Frees the alternate stack a little sooner during thread termination.
