+++ src/pal/src/exception/signal.cpp
@@ -70,15 +70,6 @@ typedef void *siginfo_t;
 #endif  /* !HAVE_SIGINFO_T */
 typedef void (*SIGFUNC)(int, siginfo_t *, void *);
 
-#if !HAVE_MACH_EXCEPTIONS
-// Return context and status for the signal_handler_worker.
-struct SignalHandlerWorkerReturnPoint
-{
-    bool returnFromHandler;
-    CONTEXT context;
-};
-#endif // !HAVE_MACH_EXCEPTIONS
-
 /* internal function declarations *********************************************/
 
 static void sigterm_handler(int code, siginfo_t *siginfo, void *context);
@@ -104,9 +95,9 @@ static void restore_signal(int signal_id, struct sigaction *previousAction);
 /* internal data declarations *********************************************/
 
 #if !HAVE_MACH_EXCEPTIONS
-static bool registered_signal_handlers = false;
+bool g_registered_signal_handlers = false;
 #endif // !HAVE_MACH_EXCEPTIONS
-static bool registered_sigterm_handler = false;
+static bool g_registered_sigterm_handler = false;
 
 struct sigaction g_previous_sigterm;
 #if !HAVE_MACH_EXCEPTIONS
@@ -129,101 +120,6 @@ int g_common_signal_handler_context_locvar_offset = 0;
 
 /* public function definitions ************************************************/
 
-#if !HAVE_MACH_EXCEPTIONS
-/*++
-Function :
-    EnsureSignalAlternateStack
-
-    Ensure that alternate stack for signal handling is allocated for the current thread
-
-Parameters :
-    None
-
-Return :
-    TRUE in case of a success, FALSE otherwise
---*/
-BOOL EnsureSignalAlternateStack()
-{
-    int st = 0;
-
-    if (registered_signal_handlers)
-    {
-        stack_t oss;
-
-        // Query the current alternate signal stack
-        st = sigaltstack(NULL, &oss);
-
-        if ((st == 0) && (oss.ss_flags == SS_DISABLE))
-        {
-            // There is no alternate stack for SIGSEGV handling installed yet so allocate one
-
-            // We include the size of the SignalHandlerWorkerReturnPoint in the alternate stack size since the 
-            // context contained in it is large and the SIGSTKSZ was not sufficient on ARM64 during testing.
-            int altStackSize = SIGSTKSZ + ALIGN_UP(sizeof(SignalHandlerWorkerReturnPoint), 16) + GetVirtualPageSize();
-#ifdef HAS_ASAN
-            // Asan also uses alternate stack so we increase its size on the SIGSTKSZ * 4 that enough for asan
-            // (see kAltStackSize in compiler-rt/lib/sanitizer_common/sanitizer_posix_libcdep.cc)
-            altStackSize += SIGSTKSZ * 4;
-#endif
-            altStackSize = ALIGN_UP(altStackSize, GetVirtualPageSize());
-            void* altStack = mmap(NULL, altStackSize, PROT_READ | PROT_WRITE, MAP_ANONYMOUS | MAP_STACK | MAP_PRIVATE, -1, 0);
-            if (altStack != MAP_FAILED)
-            {
-                // create a guard page for the alternate stack
-                st = mprotect(altStack, GetVirtualPageSize(), PROT_NONE);
-                if (st == 0)
-                {
-                    stack_t ss;
-                    ss.ss_sp = (char*)altStack;
-                    ss.ss_size = altStackSize;
-                    ss.ss_flags = 0;
-                    st = sigaltstack(&ss, NULL);
-                }
-
-                if (st != 0)
-                {
-                    int st2 = munmap(altStack, altStackSize);
-                    _ASSERTE(st2 == 0);
-                }
-            }
-        }
-    }
-
-    return (st == 0);
-}
-
-/*++
-Function :
-    FreeSignalAlternateStack
-
-    Free alternate stack for signal handling
-
-Parameters :
-    None
-
-Return :
-    None
---*/
-void FreeSignalAlternateStack()
-{
-    if (registered_signal_handlers)
-    {
-        stack_t ss, oss;
-        // The man page for sigaltstack says that when the ss.ss_flags is set to SS_DISABLE,
-        // all other ss fields are ignored. However, MUSL implementation checks that the 
-        // ss_size is >= MINSIGSTKSZ even in this case.
-        ss.ss_size = MINSIGSTKSZ;
-        ss.ss_flags = SS_DISABLE;
-        int st = sigaltstack(&ss, &oss);
-        if ((st == 0) && (oss.ss_flags != SS_DISABLE))
-        {
-            int st = munmap(oss.ss_sp, oss.ss_size);
-            _ASSERTE(st == 0);
-        }
-    }
-}
-#endif // !HAVE_MACH_EXCEPTIONS
-
 /*++
 Function :
     SEHInitializeSignals
@@ -236,14 +132,14 @@ Parameters :
 Return :
     TRUE in case of a success, FALSE otherwise
 --*/
-BOOL SEHInitializeSignals(DWORD flags)
+BOOL SEHInitializeSignals(CorUnix::CPalThread *pthrCurrent, DWORD flags)
 {
     TRACE("Initializing signal handlers\n");
 
 #if !HAVE_MACH_EXCEPTIONS
     if (flags & PAL_INITIALIZE_REGISTER_SIGNALS)
     {
-        registered_signal_handlers = true;
+        g_registered_signal_handlers = true;
 
         /* we call handle_signal for every possible signal, even
            if we don't provide a signal handler.
@@ -273,7 +169,7 @@ BOOL SEHInitializeSignals(DWORD flags)
 #ifdef INJECT_ACTIVATION_SIGNAL
         handle_signal(INJECT_ACTIVATION_SIGNAL, inject_activation_handler, &g_previous_activation);
 #endif
-        if (!EnsureSignalAlternateStack())
+        if (!pthrCurrent->EnsureSignalAlternateStack())
         {
             return FALSE;
         }
@@ -292,7 +188,7 @@ BOOL SEHInitializeSignals(DWORD flags)
 
     if (flags & PAL_INITIALIZE_REGISTER_SIGTERM_HANDLER)
     {
-        registered_sigterm_handler = true;
+        g_registered_sigterm_handler = true;
         handle_signal(SIGTERM, sigterm_handler, &g_previous_sigterm);
     }
 
@@ -320,7 +216,7 @@ void SEHCleanupSignals()
     TRACE("Restoring default signal handlers\n");
 
 #if !HAVE_MACH_EXCEPTIONS
-    if (registered_signal_handlers)
+    if (g_registered_signal_handlers)
     {
         restore_signal(SIGILL, &g_previous_sigill);
         restore_signal(SIGTRAP, &g_previous_sigtrap);
@@ -335,7 +231,7 @@ void SEHCleanupSignals()
     }
 #endif // !HAVE_MACH_EXCEPTIONS
 
-    if (registered_sigterm_handler)
+    if (g_registered_sigterm_handler)
     {
         restore_signal(SIGTERM, &g_previous_sigterm);
     }
